refactor(profile): merge layout
